Only allow env vars to add labels
