+++ metaflow/plugins/kubernetes/kubernetes_cli.py
@@ -7,10 +7,10 @@ from metaflow import JSONTypeClass, util
 from metaflow._vendor import click
 from metaflow.exception import METAFLOW_EXIT_DISALLOW_RETRY, CommandException
 from metaflow.metadata.util import sync_local_metadata_from_datastore
-from metaflow.metaflow_config import DATASTORE_LOCAL_DIR
+from metaflow.metaflow_config import DATASTORE_LOCAL_DIR, KUBERNETES_LABELS
 from metaflow.mflog import TASK_LOG_SOURCE
 
-from .kubernetes import Kubernetes, KubernetesKilledException
+from .kubernetes import Kubernetes, KubernetesKilledException, parse_kube_keyvalue_list
 from .kubernetes_decorator import KubernetesDecorator
 
 
@@ -97,12 +97,6 @@ def kubernetes():
     type=JSONTypeClass(),
     multiple=False,
 )
-@click.option(
-    "--labels",
-    multiple=True,
-    default=None,
-    help="Labels for Kubernetes pod.",
-)
 @click.pass_context
 def step(
     ctx,
@@ -126,7 +120,6 @@ def step(
     tmpfs_path=None,
     run_time_limit=None,
     tolerations=None,
-    labels=None,
     **kwargs
 ):
     def echo(msg, stream="stderr", job_id=None, **kwargs):
@@ -192,12 +185,7 @@ def step(
     stderr_location = ds.get_log_location(TASK_LOG_SOURCE, "stderr")
 
     # `node_selector` is a tuple of strings, convert it to a dictionary
-    node_selector = KubernetesDecorator.parse_kube_keyvalue_list(node_selector)
-
-    # `labels` is a tuple of strings or a tuple with a single comma separated string
-    # convert it to a dict
-    labels = KubernetesDecorator.parse_kube_keyvalue_list(labels, False)
-    KubernetesDecorator.validate_kube_labels(labels)
+    node_selector = parse_kube_keyvalue_list(node_selector)
 
     def _sync_metadata():
         if ctx.obj.metadata.TYPE == "local":
@@ -244,7 +232,6 @@ def step(
                 run_time_limit=run_time_limit,
                 env=env,
                 tolerations=tolerations,
-                labels=labels,
             )
     except Exception as e:
         traceback.print_exc(chain=False)
